Add jammy-yoga distro regression jobs.
There is an agreement between the OpenStack and Ubuntu SRU team
allowing SRU of upstream point release updates on the back of
verification with the OpenStack product.

The openvswitch and ovn packages are covered by this agreement
and as such we need an automated way of performing these tests.

For jammy-yoga, the readily available way of performing this
verification is through the use of the Zaza test runner and
Charmed OpenStack bundle fragments.

Verification of more recent versions will likely be performed
using the openstack snap and its sunbeam tool and is subject for
a future PR.

Signed-off-by: Frode Nordahl <fnordahl@ubuntu.com>