Add spread submodule.
The spread project is an integral part of other Canonical projects
so the core of it is well maintained.  It has a light weight build
process so building and installing it as part of a CI workflow is
trivial.

We require both VMs and up to date images for our tests, and LXD
provides this.  The spread LXD backend currently does not support
LXD VMs, but there are a couple of efforts in-flight to add this
support.

While maintained, the spread project is slow moving, so there is
little risk to making use of in-flight patches ahead of time.

Use fork of spread with LXD VM support.

Signed-off-by: Frode Nordahl <fnordahl@ubuntu.com>