CI: OVN and OVS scheduled coverage tests.
This PR adds automatic CI workflows that run every Tuesday at 02:21,
and execute unit tests with coverage on the `main` branch of OVN and
OVS repository. Results of the coverage run are uploaded to their
respective TIOBE TICS projects.

Logic to get the repositories, install dependencies, configure,
build and test the projects is implemented in the `justfile`. For
more info see README.md

Signed-off-by: Martin Kalcok <martin.kalcok@canonical.com>